requirements_pyHechtPlayer.txt
pyautogui
schedule
playsound
dbus-python
tk
